Move Geofabrik file handling to dedicated module. Additional Python tests
